Phase 3 (suite): resoudre C4244 restants + C4701/C4703, code AzurVeda a 0 avertissement /W4
Elimine les derniers avertissements C4244 (conversion avec perte de
donnees) restants dans le code propre a AzurVeda : ajout de casts
explicites (unsigned char)/(signed char)/(float) aux points de
troncature intentionnelle (quantification 0-255, serialisation
d'octets, mise a l'echelle de temps double -> float d'affichage).

Vrai bug corrige : dans EquationSinus3D.cpp, les champs
mSer_IndexMultiplier_Frequence et mSer_IndexMultiplier_Amplitude sont
des PackDynamicFloat mais etaient enregistres via la macro
REGISTER_MEMBER_PACKFLOAT au lieu de REGISTER_MEMBER_PACKDYNAMICFLOAT.
Celle-ci appelle Set(_DefaultVal), qui pour un PackDynamicFloat
resout vers PackDynamicType::Set(unsigned int) -- une methode sans
rapport -- et corrompait silencieusement l'etat "shape" au lieu de
fixer une valeur par defaut flottante. Recherche exhaustive (agent
Explore dedie) confirmant qu'il s'agit du seul cas de ce pattern dans
la base de code.

Traite ensuite les avertissements C4701/C4703 (variable/pointeur
potentiellement non initialise) restants, invisibles a la premiere
passe car MSVC les emet au format "(ligne) : warning" (avec espace)
plutot que "(ligne): warning" comme les autres diagnostics -- le
script d'analyse a ete corrige en consequence.

Deux vrais bugs corriges parmi ces cas :
- BaseType.cpp::WriteFile : pserializedFormAfter n'etait jamais
  affecte si fopen() echouait, et sa valeur non initialisee etait
  lue juste apres. Le calcul de verification est deplace a
  l'interieur du bloc ou fopen() a reussi.
- lwo2reader.c : plwclip restait non initialise si nbclip==0, mais un
  chunk CLIP present dans un fichier LWO incoherent y aurait quand
  meme deref au travers d'un pointeur non initialise. Initialisation
  a NULL et garde ajoutee.

Les autres cas (RGBAImage.cpp, Object3DMesh3D.cpp,
Object3DMarchCube.cpp) sont des faux positifs de l'analyse de flux du
compilateur (les variables sont toujours affectees avant usage reel
compte tenu de la logique de garde existante) ; initialisees a des
valeurs par defaut sures par prudence, sans changement de
comportement sur les chemins de donnees valides.

Recompilation complete verifiee : 0 erreur, 0 avertissement /W4 dans
le code AzurVeda (Veda, VedaLib*, VedaGUIWindowsMFC,
VedaMachineOGL*). Seuls les 68 avertissements du code vendored
MP3/XM (libmad, uniminixm) subsistent, prochaine etape avant
activation de /WX.

+++ Veda/CodeExamples/E02_SimpleMachineUse/SimpleMachineUse.cpp
@@ -20,7 +20,7 @@ class miniSoundGenerator : public VirtualMachine::SoundInterface
 
 		float freq = _SoundBufferToAddYourSignal.m_PlayFrequency ;
 		float speed1 =  440.0f*4.0f/freq;
-		float speed2 =  293.3*4.0f/freq;
+		float speed2 =  293.3f*4.0f/freq;
 
 		unsigned int ii;
 		float fp1 = m_sinusPostion1;
@@ -30,10 +30,10 @@ class miniSoundGenerator : public VirtualMachine::SoundInterface
 
 		for(ii=0 ; ii<lengthToRender ; ii++)
 		{
-			*(pSoundBuffer++) += sin( fp1 ) *leftVolume ;
-			*(pSoundBuffer++) += sin( fp2 ) *rightVolume;
-			fp1 +=(speed1 + sin(fp2*0.0018f)*0.004f);
-			fp2 += speed2+ sin(fp1*0.002f)*0.004f;
+			*(pSoundBuffer++) += sinf( fp1 ) *leftVolume ;
+			*(pSoundBuffer++) += sinf( fp2 ) *rightVolume;
+			fp1 +=(speed1 + sinf(fp2*0.0018f)*0.004f);
+			fp2 += speed2+ sinf(fp1*0.002f)*0.004f;
 		}
 
 		m_sinusPostion1 = fp1;
